Move molecule into separate component
